+++ src/pptx/oxml/ns.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 _nsmap = {
     "a": "http://schemas.openxmlformats.org/drawingml/2006/main",
     "c": "http://schemas.openxmlformats.org/drawingml/2006/chart",
+    "c14": "http://schemas.microsoft.com/office/drawing/2007/8/2/chart",
     "cp": "http://schemas.openxmlformats.org/package/2006/metadata/core-properties",
     "ct": "http://schemas.openxmlformats.org/package/2006/content-types",
     "dc": "http://purl.org/dc/elements/1.1/",